% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2MCDS-package.r
\docType{data}
\name{alcidae}
\alias{alcidae}
\title{sample dataset distance analysis}
\format{A data frame with 1670 observations on the following 14 variables.
 \describe{
   \item{\code{CruiseID}}{unique cruise ID.}
   \item{\code{WatchID}}{unique watch ID.}
   \item{\code{Observer}}{Categorical variable identifying the different observers.}
   \item{\code{Date}}{date of the watch.}
   \item{\code{StartTime}}{start time of the watch.}
   \item{\code{EndTime}}{end time of the watch.}
   \item{\code{LatStart}}{latitude at start of the watch, decimal degrees.}
   \item{\code{LongStart}}{longitude at start of the watch, decimal degrees.}
   \item{\code{WatchLenKm}}{distance travelled during the watch in kilometers (estimated from platform speed and observation length since many watches don't have start and end positions).}
   \item{\code{Alpha}}{four letter code of the species.}
   \item{\code{English}}{English name.}
   \item{\code{Latin}}{scientific name.}
   \item{\code{Distance}}{Classes of distances for the observations. A=0-50m, B=50-100m, C=100-200m, D=200-300m.}
   \item{\code{Count}}{number of individuals.}
 }}
\description{
A sample dataset containing observation of alcidae made in the Eastern Canada Seabirds At Sea (ECSAS) program.
}
\seealso{
\code{\link{mcds.filter}}
}
