% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create.grid.R
\name{create.grid}
\alias{create.grid}
\title{Create a grid using latitude and longitude values.}
\usage{
create.grid(Latitude = c(45, 52), Longitude = c(-70, -56),
  Grid.size = c(0.1, 0.1), Clip = FALSE, clip.shape, projection = NULL)
}
\arguments{
\item{Latitude}{A pair of coordinates that identifies the grid northern and southern limits.}

\item{Longitude}{A pair of coordinates that identifies the grid eastern and western limits.}

\item{Grid.size}{size of the grid cells in km.}

\item{Clip}{Should the grid be clipped with a shapefile (\env{TRUE}) or not (\env{FALSE}).}

\item{clip.shape}{Shapefile to use to clip the grid if Clip is set to \env{TRUE}}

\item{projection}{coordinate reference system ID for the grid.}

\item{x}{A \code{\link{data.frame}} containing observations.}
}
\description{
This function creates a \code{\link{SpatialPolygonsDataFrame}} grid using the boundaries and cell size selected by the user.
}
\details{
The projection used for the grid must match the projection used for the dataset if you want to use the grid for \code{\link{effort.grid}} and \code{\link{species.grid}}.
}
\section{Author}{
Christian Roy
}

\section{note}{
Large grid with small cell size can take a while to create. Be patient.
}
\examples{
new.grid<-create.grid(Latitude=c(45, 54), Longitude=c(-70, -56), Grid.size=c(25000, 25000), Clip=FALSE, clip.shape=canshp, projection=NULL)
#plot the grid
plot(new.grid, axes=T)
#END
}

