% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{create.maps}
\alias{create.maps}
\title{Plot the density estimates for a given dataset and grid.}
\usage{
create.maps(Observations.grid, Observations, only.visits = TRUE,
  Land = FALSE, Land.shapefile, Title = NULL)
}
\arguments{
\item{Observations.grid}{output of a \code{\link{species.grid}}}

\item{Observations}{which column of the Observation.grid should be used to plot abundance}

\item{only.visits}{\env{TRUE} if the user want only the cells of the grid that have been visited. \env{FALSE} will plot all the cells of the grid.}

\item{Land}{\env{TRUE} if the user want to plot an additional shapefile over the grid. \env{FALSE} if the user want only the grid to be plotted.}

\item{Land.shapefile}{name of the object the user want to plot over the grid. Must be an object of the class \code{\link{SpatialPolygonsDataFrame}}}

\item{Title}{Title for the figure. Must be a \code{\link{character}} vector.}
}
\description{
This function plots output of the \code{\link{species.grid}} function. The user has the option to plot only the cells that have been visited or all the cells.
}
\details{
This function return a plot with the numbers of observations per cell and a scale to help visualisation. The upper limit for the scale is 19.
}
\section{Author}{
Christian Roy
}
\examples{
data(alcidae)
###Create  a grid
new.grid<-create.grid(Latitude=c(45, 54), Longitude=c(-70, -56), Grid.size=c(25000, 25000), Clip=FALSE, projection=NULL)
###calculate the density of alcidea for the grid
alcidae.grid<-species.grid(alcidae, Grid.shp=new.grid, Selection="Alpha",
                     Code=c("ALCI"), Density="Count",Latitude="LatStart",Longitude="LongStart",
                     Cell.id="CELL.NUM", Factors=F)
#plot a map with only the cells that have been visited
create.maps(Observations.grid=alcidae.grid, Observations="Count", only.visits=TRUE, Land=FALSE)
#plot a map with all the cells
create.maps(Observations.grid=alcidae.grid, Observations="Count", only.visits=FALSE, Land=FALSE)
#use the create.maps function to visualize effort in each grid cell.
create.maps(Observations.grid=alcidae.grid, Observations="EFFORT", only.visits=FALSE, Land=FALSE, Title="Number of time a cell has been visited")
#END
}

