% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/distance.wrap.R
\name{distance.wrap}
\alias{distance.wrap}
\title{Fit the distance sampling model provided by the Distance software.}
\usage{
distance.wrap(dataset, path, pathMCDS, SMP_LABEL = "SMP_LABEL",
  SMP_EFFORT = "SMP_EFFORT", DISTANCE = "DISTANCE", SIZE = "SIZE",
  STR_LABEL = "STR_LABEL", STR_AREA = "STR_AREA", units = list(Type =
  "Line", Distance = "Perp", Length_units = "Kilometer", Distance_units =
  "Meter", Area_units = "Square kilometer"), breaks = c(0, 50, 100, 200, 300),
  covariates = NULL, factor = NULL, lsub = NULL, stratum = NULL,
  split = TRUE, rare = NULL, period = NULL, detection = "All",
  monotone = "Strict", estimator = NULL, multiplier = 2, empty = NULL,
  verbose = FALSE)
}
\arguments{
\item{dataset}{A \code{\link{data.frame}} with observations.}

\item{path}{The path where to store the input and output files of the MCDS engine.}

\item{pathMCDS}{The path where the MCDS engine is located.}

\item{SMP_LABEL}{Name of the column to use for the transect/watch label.}

\item{SMP_EFFORT}{Length in of the transect or the transect/watch unit.}

\item{DISTANCE}{Distance of the observation.}

\item{SIZE}{Number of individuals in the observation.}

\item{STR_LABEL}{Name of the column to use as a stratum label.}

\item{STR_AREA}{Name of the column to use for the stratum area.}

\item{units}{list of th units used for the analysis. Contains the Type of analysis (Line or Point), the Distance enfine to use (Perp or Radial),
the Length units, the Distance units, and the Area_units.For the possible units of distance and area see Distance 6.2 documentation.}

\item{breaks}{A vector giving the distance intervals in meters to be used in the analysis.}

\item{covariates}{A vector giving the name of covariates.}

\item{factor}{A vector giving the name of factors to be used in the analysis.}

\item{lsub}{A named list giving the subsets to be used in the analysis. The names of the list are the names of columns used to subset the dataset.
Each element of the list is a vector giving the values to keep in the analysis for a given column. When a vector is \code{NULL},
all values are kept for this column. Default value \code{lsub = NULL}. See examples for further details.}

\item{stratum}{When \code{stratum = "STR_LABEL"}, the model will be stratified with the label.
When \code{stratum} is the name of a column in the data, the model will be
post-stratified according to this column (see Thomas et al. 2010). Default value
is \code{NULL}.}

\item{split}{When \code{split = TRUE}, separate models are ran according to the subsets determined by \code{lsub}. Default value \code{FALSE}.}

\item{rare}{This argument is used when a species has few observations to estimate a detection function.
The probability of detection is estimated from a group of similar species and a multiplier. A named list
has to be given, with the column name where species are stored as the name of the list and the name of the species as
an element (ex: \code{list(Species = "HERG")}). Default value \code{NULL}. See details.}

\item{period}{A vector of characters of length 2 containing the extreme dates for which the analysis
should be restricted. Dates have to be in the "yyyy-mm-dd" format.}

\item{detection}{Currently, set to \code{detection = "All"}. Can also be \code{detection = "Stratum"}.}

\item{monotone}{Currently, set to \code{monotone = "Strict"}.}

\item{estimator}{When set to \code{NULL}, the following key functions and expansion terms will be used: UN-CO, UN-PO, HN-CO, HN-HE, HA-CO and HA-PO.
If the user wants to choose the key functions and the expansion terms used, a list has to be given with each element a vector of
length 2 with the first element the key function and the second element the expansion term (ex: \code{list(c("HN","CO"),c("HA","PO")}).
When \code{rare != NULL}, only one set is used (UN-CO) for the final specific model.}

\item{multiplier}{Value by which the estimates of density or abundance are multiplied. Default value \code{multiplier = 2} meaning
only one-half of the transect is surveyed. When \code{rare != NULL}, the multiplier will be modified to account
for the probability of detection.}

\item{empty}{Determine how empty transects are to be selected in the analysis. When \code{empty = NULL}, all
empty transects are included for every element of \code{lsub}. For example, when models are splitted according
to species, empty transects and transects where a species was not detected need to be considered in the analysis for that species.
When \code{lsub} contains geographic or temporal elements, empty transects need to be restricted to the subsets given. In this case,
a vector of character has to be given with the names in \code{lsub} for which the empty transects are to be restricted. When \code{split = TRUE}
and \code{empty != NULL}, empty transects will be splitted according to the names in \code{empty}. In any case, it is assumed that
the \code{dataset} contains at least a line for every transect executed, either with or without an observation. See examples for
further details.}

\item{verbose}{When set to \code{TRUE}, prints the input file given to the MCDS engine. Default value \code{FALSE}.}
}
\value{
An object of class \code{"distanceFit"}, when \code{split = FALSE}. When \code{split = TRUE} and \code{lsub != NULL},
a named list with the different models of class \code{"distanceFit"}.
Each object of class \code{"distanceFit"} is a named list with components \code{model_fitting}, \code{parameter_estimates},
\code{chi_square_test}, \code{density_estimate}, \code{detection} and \code{path}. Elements
of the list are accessible through the \code{$} operator. For each component
except \code{detection} and \code{path}, a \code{\link{list}} of length 2 is given with
component "\code{Global}" and "\code{Stratum}". Depending on the analysis chosen,
one of these component will be empty (\code{NULL}).
}
\description{
This function fits the distance sampling model provided by the Distance program using its MCDS engine.
}
\details{
Uses the MCDS engine from program Distance. The function produces an input file and submits it to the MCDS engine
through the \code{\link{system}} function. The results are then extracted from the output files and returned as a
list object.
}
\section{Author}{
Francois Rousseu, Christian Roy, Francois Bolduc
}
\examples{
########################################
### Simple models without stratification
### Import and filter data
data(alcidae)
alcids<-filterECSAS(alcidae)

### Run analysis with the MCDS engine. Here, the WatchID is used as the sample.
dist.out1 <-distance.wrap(alcids, SMP_EFFORT="WatchLenKm",DISTANCE="Distance",SIZE="Count",
                         units=list(Type="Line",Distance="Perp",Length_units="Kilometers",
                                    Distance_units="Meters",Area_units="Square kilometers"),
                         breaks=c(0,50,100,200,300), estimator=list(c("HN","CO")),
                         STR_LABEL="STR_LABEL", STR_AREA="STR_AREA",SMP_LABEL="WatchID",
                         path="c:/temp/distance",
                         pathMCDS="C:/Program Files (x86)/Distance 6")

summary(dist.out1)
global.summary(model=dist.out1, species="alcidae",file="alcidae_global", directory="C:/temp/distance")

### Run separate analysis for years 2008-2009
dist.out2 <- distance.wrap(alcids, SMP_EFFORT="WatchLenKm",DISTANCE="Distance",SIZE="Count",
                          units=list(Type="Line",Distance="Perp",Length_units="Kilometers",
                                     Distance_units="Meters",Area_units="Square kilometers"),
                          breaks=c(0,50,100,200,300), estimator=list(c("HN","CO")),
                          lsub=list(Year=c(2008,2009)), split=TRUE, empty="Year",
                          STR_AREA="STR_AREA",SMP_LABEL="WatchID",
                          path="c:/temp/distance",
                          pathMCDS="C:/Program Files (x86)/Distance 6")

### Get the names of the different models produced
names(dist.out2)
#####summary for the Year 2008 model
summary(dist.out2[["2008"]])
#### Plot results for the Year2008 model
global.summary(model=dist.out2[["2008"]], species="alcidae", file="alcidae", directory="C:/temp/distance")
#END
}
\references{
Thomas, L., S.T. Buckland, E.A. Rexstad, J. L. Laake, S. Strindberg, S. L. Hedley, J. R.B. Bishop,
T. A. Marques, and K. P. Burnham. 2010. \emph{Distance software: design and analysis of distance sampling
surveys for estimating population size.} Journal of Applied Ecology 47: 5-14.
DOI: 10.1111/j.1365-2664.2009.01737.x
}

