% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/effort.grid.R
\name{effort.grid}
\alias{effort.grid}
\title{Calculate the effort for each grid cell.}
\usage{
effort.grid(Observations = alcidae, Grid.shp = new.grid,
  Latitude = "LatStart", Longitude = "LongStart", TransectID = "WatchID",
  CellID = "CELL.NUM")
}
\arguments{
\item{Observations}{data.frame with the observations.}

\item{Grid.shp}{a \code{\link{SpatialPolygonsDataFrame}} grid that is overlapping with the observations in the data.frame \cr \emph{une grille de format SpatialPolygonsDataFrame qui chevauche les observations dans le data.frame}}

\item{Latitude}{which column of the data.frame hold the latitude for the observation}

\item{Longitude}{which column of the data.frame hold the longitude for the observation}

\item{TransectID}{which colum of the data.frame hold the ID for each independant transect.}

\item{CellID}{which colum identify each cell individually in the a \code{\link{SpatialPolygonsDataFrame}} grid}
}
\value{
\item{x}{
longitude value of the centroid.
}
 \item{y}{
latitude value of the centroid.
}
 \item{EFFORT}{
number of time the cell has been visited.
}
}
\description{
This function will calculate the effort (i.e the number of time it was visited) for each cells of a grid given a dataset with spatial references.
}
\details{
This function will return the grid appended with new values for each cells.
}
\section{Author}{
Christian Roy
}
\examples{
data(quebec)
x<-distance.filter(quebec, transect.id = "WatchID", distance.field = "Distance", distance.labels = c("A", "B", "C", "D"),
                  distance.midpoints = c(25, 75, 150, 250), effort.field = "WatchLenKm", lat.field = "LatStart",
                  long.field = "LonStart", sp.field = "Alpha", date.field = "Date")
str(x)
END
}
\seealso{
\code{\link{species.grid}}
\code{\link{create.maps}}
}

