% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/filterECSAS.R
\name{filterECSAS}
\alias{filterECSAS}
\title{Filter data collected according to the Eastern Canada Seabirds At Sea (ECSAS) standardized protocol for pelagic seabird surveys.}
\usage{
filterECSAS(x, dist2m = TRUE)
}
\arguments{
\item{x}{A \code{\link{data.frame}} containing observations.}

\item{dist2m}{Logical value. Should the distance classes be transformed in meters?}
}
\description{
Prepares the raw data gathered by the ECSAS for a distance analysis.
}
\details{
The following column must be in \code{x}: "WatchLenKm","InTransect","Distance" and "Date".
When "WatchLenKm" = 0, observations are eliminated. Only observations within the transect ("InTransect" = 1) are kept. Transects for
which "Alpha" = "" (no species names) will also be kept because they are transects that were done but where no observations were recorded.
Only distance class "A", "B", "C" and "D" will be kept and if \code{dist2m = TRUE},
they will be transformed to 25, 75, 150 and 250, respectively. Missing distance will also be kept if the observation corresponds to an empty transect.
Observations for which there is no distance or coordinates will be eliminated. The "Date" column will be transformed in the yyyy-mm-dd format.
}
\section{Author}{
Francois Rousseu
}
\examples{
data(quebec)
x<-filterECSAS(quebec)
str(x)
#End
}
\seealso{
\code{\link{alcidae}}, \code{\link{quebec}}
}

