% Generated by roxygen2 (4.0.0.99): do not edit by hand
\docType{package}
\name{GeoAviR}
\alias{GeoAviR}
\alias{GeoAviR-package}
\title{Creating spatial grids and implementing distance sampling from the package unmarked.}
\description{
This package allows the user to create spatial grids, plot the animal density and run some basic distance analysis on the dataset.
}
\details{
\tabular{ll}{
Package: \tab GeoAviR\cr
Type: \tab Package\cr
Version: \tab 0.6.5\cr
Date: \tab 2014-05-17\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
}
\examples{
data(alcidae)
hist.wrap(alcidae,Trn.Value="Count", Dist.class="Distance", Keep.class=c("A", "B", "C", "D"),
         Breaks=c(0,50,100,200,300), Trn.ID="WatchID",  Trn.length="WatchLenKm", Resume=FALSE)
alcidae.output<-unmarked.wrap(alcidae,Trn.Value="Count", Dist.class="Distance", Keep.class=c("A", "B", "C", "D"),
                            Breaks=c(0,50,100,200,300),  Trn.ID="WatchID",  Trn.length="WatchLenKm", Resume=TRUE)
str(alcidae.output)
summary(alcidae.output[[2]])
par(ask=T)
resume.plot(alcidae.output[[1]])
resume.plot(alcidae.output[[2]])
resume.plot(alcidae.output[[3]])
par(ask=F)
}
\references{
S.T. Buckland, D.R. Anderson, K.P. Burnham, J.L Laake, D.L. Borchers and L. Thomas. 2001. \emph{Introduction to Distance Sampling}. Estimating abundance of biological populations. Oxford University Press.\cr
Royle, J. A., D. K. Dawson, and S. Bates. 2004. \emph{Modeling abundance effects in distance sampling}. Ecology 85:1591–1597\cr
Fiske, I. and R. B. Chandler. 2011. \emph{unmarked: An R package for fitting hierarchical models of wildlife occurrence and abundance}. Journal of Statistical Software 43:1–23.\cr
Thomas, L., S.T. Buckland, E.A. Rexstad, J. L. Laake, S. Strindberg, S. L. Hedley, J. R.B. Bishop, T. A. Marques, and K. P. Burnham. 2010.  Distance software: design and analysis of distance sampling surveys for estimating population size.  Journal of Applied Ecology 47: 5-14.
}

