% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{hist.wrap}
\alias{hist.wrap}
\title{Plot a detection histogram of apparent density.}
\usage{
\method{hist}{wrap}(Dataset, Trn.Value, Dist.class, Keep.class, Breaks, Trn.ID,
  Trn.length, Resume = F)
}
\arguments{
\item{x}{A \code{\link{data.frame}} containing observations.}

\item{Dataset}{what is the name of the dataset that will be used for the analysis.}

\item{Trn.Value}{which column in the data.frame hold the information for the abundance.}

\item{Dist.class}{which column in the data.frame hold the information for the distance class.}

\item{Keep.class}{which class of observations should be used for the analysis.}

\item{Breaks}{what are the intervals for the distance analysis.}

\item{Trn.ID}{which column in the data.frame hold the unique ID for each transect.}

\item{Trn.length}{which column in the data.frame hold the length of each transect.}

\item{Resume}{if \env{Resume} is set to \env{TRUE} the analysis will be performed as one giant transect instead of several small transect.}
}
\description{
This function will allow the user to visualise the histogram of observations in function of distance class and re-bin his observations if necessary before the analysis.
}
\details{
When "WatchLenKm" = 0, observations are eliminated. Transects for which "Alpha" = "" (no species names)
will be kept because they are transects that were done but where no observations were recorded.
Observations for which there is no distance or coordinates will be eliminated.
The "Date" column will be transformed in the yyyy-mm-dd format.
}
\section{Author}{
Christian Roy
}
\examples{
###Import the data
data(alcidae)

###Check the naïve detection histogram
hist.wrap(alcidae,Trn.Value="Count", Dist.class="Distance", Keep.class=c("A", "B", "C", "D"),
         Breaks=c(0,50,100,200,300), Trn.ID="WatchID",  Trn.length="WatchLenKm", Resume=F)
#END
}
\seealso{
\code{\link{unmarked.wrap}}
}

