% Generated by roxygen2 (4.0.2): do not edit by hand
\name{keep.best.model}
\alias{keep.best.model}
\title{Keep the best model of a distanceList object}
\usage{
keep.best.model(x)
}
\arguments{
\item{x}{A \code{\link{distanceList}} object.}
}
\value{
An object of class \code{"distanceFit"}
}
\description{
This function keep the model with the lowest AICc value in a distanceList object.
}
\details{
Select the model with the smallest AICc value. If more than one model have a Delta AICc equal to zero the uniform models
are discaded and the best model is chosen randomly between the remaining models.
}
\section{Author}{
Christian Roy
}

