% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcds.filter.R
\name{mcds.filter}
\alias{mcds.filter}
\title{Filter data before using it with mcds.wrap}
\usage{
mcds.filter(x, transect.id = "WatchID", distance.field = "Distance",
  distance.labels = c("A", "B", "C", "D"), distance.midpoints = c(25,
  75, 150, 250), effort.field = "WatchLenKm", lat.field = "LatStart",
  long.field = "LongStart", sp.field = "Alpha", date.field = "Date",
  distanceLabel.field = "Distance", dist2m = TRUE)
}
\arguments{
\item{x}{A \code{\link{data.frame}} containing observations.}

\item{transect.id}{Name of the column containing the unique ID of each transect.}

\item{distance.field}{Name of the column containing the distance classes of each observations.}

\item{distance.labels}{Classes of distance to keep for the analysis.}

\item{distance.midpoints}{Midpoints in (m) of the classes of distance kept for the analysis.}

\item{effort.field}{Name of the column containing the length of the transect/watch.}

\item{lat.field}{Name of the column containing the latitude of the observations.}

\item{long.field}{Name of the column containing the longitude of the observations.}

\item{sp.field}{Name of the column containing the species ID.}

\item{date.field}{Name of the column containing the date for the observations.}

\item{dist2m}{Boolean to indicate if the conversion from classes to numeric should be performed.}

\item{distanceLabels.field}{Name of the column containing the distance classes for filtering.}
}
\description{
Allow to change the names of the column of a data.frame so that it work seamlessly with \code{\link{mcds.wrap}}
}
\details{
When "WatchLenKm" = 0, observations are eliminated. Transects for which "Alpha" = "" (no species names) 
will be kept because they are transects that were done but where no observations were recorded.
Observations for which there is no distance or coordinates will be eliminated. 
The "Date" column will be transformed in the yyyy-mm-dd format.
}
\section{Author}{
Christian Roy
}

\examples{
data(quebec)
x<- mcds.filter(quebec, transect.id = "WatchID", distance.field = "Distance", distance.labels = c("A", "B", "C", "D"), 
                  distance.midpoints = c(25, 75, 150, 250), effort.field = "WatchLenKm", lat.field = "LatStart", 
                  long.field = "LongStart", sp.field = "Alpha", date.field = "Date")
str(x)
}
