% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation_hist.R
\name{observation_hist}
\alias{observation_hist}
\title{Plot a detection histogram of apparent density.}
\usage{
observation_hist(dataset, count, dist.class, keep.class, breaks,
  color = "white", rescale = NA)
}
\arguments{
\item{dataset}{A \code{\link{data.frame}} containing observations.}

\item{count}{which column in the data.frame hold the information for the cluster size.}

\item{dist.class}{which column in the data.frame hold the information for the distance class.}

\item{keep.class}{which class of observations should be used for the analysis.}

\item{breaks}{what are the intervals for the distance analysis.}

\item{color}{fill color for the bars}

\item{rescale}{set the height of the first bin}
}
\description{
This function will allow the user to visualise the histogram of observations in function of distance class.
}
\details{
Make a histogram of the observations in function of the distance class
}
\section{Author}{
Christian Roy
}
\examples{
###Import the data
data(alcidae)

###Check the naïve detection histogram for alcidae
observation_hist(alcidae, count="Count", dist.class="Distance", keep.class=c("A", "B", "C", "D"), breaks=c(0,50,100,200,300), color="white")
#END
}

