% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GeoAviR-package.r
\docType{data}
\name{quebec}
\alias{quebec}
\title{sample dataset for the distance analysis.}
\format{A data frame with 7360 observations on the following 19 variables.
 \describe{
   \item{\code{CruiseID}}{unique cruise ID.}
   \item{\code{Start.Date}}{start date of the cruise.}
   \item{\code{End.Date}}{end date of the cruise.}
   \item{\code{WatchID}}{unique watch ID.}
   \item{\code{Observer1}}{name of first Observer.}
   \item{\code{Date}}{date of the watch.}
   \item{\code{StartTime}}{start time of the watch.}
   \item{\code{EndTime}}{end time of the watch.}
   \item{\code{LatStart}}{latitude at start of the watch, decimal degrees.}
   \item{\code{LongStart}}{longitude at start of the watch, decimal degrees.}
   \item{\code{WatchLenKm}}{distance travelled during the watch in kilometers (estimated from platform speed and observation length since many watches don't have start and end positions).}
   \item{\code{Snapshot}}{whether snapshot method was used or not.}
   \item{\code{Alpha}}{four letter code of the species.}
   \item{\code{English}}{English name.}
   \item{\code{Latin}}{scientific name.}
   \item{\code{Distance}}{Distance to bird (in meters).}
   \item{\code{InTransect}}{Are the observations made in the transect or outside of the transect}
   \item{\code{Count}}{number of individuals.}
   \item{\code{Year}}{Year the observations were made.}
 }}
\description{
A simulated sample dataset based on the observation made in the Eastern Canada Seabirds At Sea (ECSAS) program in the Gulf of St-Lawrence.
}

