% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.grid.R
\name{species.grid}
\alias{species.grid}
\title{Calculate the density of animal in each grid cell.}
\usage{
species.grid(Observations = alcidae, Grid.shp = new.grid,
  Selection = "Alpha", Code = c("HERG"), Density = "Count",
  Latitude = "LatStart", Longitude = "LongStart", Cell.id = "CELL.NUM",
  Factors = TRUE, my.breaks = c(-1, 0, 5, 10, 15, 1000), my.labels = c(1,
  2, 3, 4, 5), ...)
}
\arguments{
\item{Observations}{data.frame holding the observations.}

\item{Grid.shp}{a \code{\link{SpatialPolygonsDataFrame}} grid that  overlapping with the obseravtions in the data.frame.}

\item{Selection}{which column of the data.frame should be used to select the observations.}

\item{Code}{which observations in the column \env{Selection} should be used.}

\item{Density}{which column of the data.frame hold the number of animals observed for a given row.}

\item{Latitude}{which column of the data.frame hold the latitude for the observation.}

\item{Longitude}{which column of the data.frame hold the longitude for the observation.}

\item{Cell.id}{which column identify each cell individually in the \code{\link{SpatialPolygonsDataFrame}} grid.}

\item{Factors}{if set to \env{TRUE} the mean density will also be reported as a class given by the \env{my.breaks} and \env{my.class} values.}

\item{my.breaks}{cut points for the class of density.}

\item{my.labels}{name of the labels for each class of density.}

\item{dots}{extra commands}
}
\value{
\item{x}{
longitude value of the centroid.
}
 \item{y}{
latitude value of the centroid.
}
 \item{EFFORT}{
number of time the cell has been visited.
}
 \item{VISITED}{
binary value. if the cell has been visited 1 otherwise 0.
}
 \item{Count}{
mean density of animal observed in the cell.
}
}
\description{
This function calculate the mean density of observation by cells of a grid given a dataset with spatial references.
}
\details{
This function returns the grid appended with new values for each cells.
}
\section{Author}{
Christian Roy
}
\examples{
###Import a dataset
data(alcidae)
###Create  a grid 
new.grid<-create.grid(Latitude=c(45, 54), Longitude=c(-70, -56), Grid.size=c(25000, 25000), Clip=F, clip.shape=canshp, projection=NULL)
###calculate the density of alcidae for the grid
alcidae.grid<-species.grid(alcidae, Grid.shp=new.grid, Selection="Alpha",
                     Code=c("ALCI"), Density="Count",Latitude="LatStart",Longitude="LongStart",
                     Cell.id="CELL.NUM", Factors=F)
#END
}

