% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_summary-methods.R
\name{print.distanceFit}
\alias{print.distanceFit}
\alias{print.distanceList}
\alias{print.summary.distanceFit}
\alias{print.summary.distanceList}
\alias{summary.distanceFit}
\alias{summary.distanceList}
\title{Summary and print methods for \code{distanceFit} and \code{distanceList} objects}
\usage{
\method{print}{distanceFit}(x, ...)

\method{print}{distanceList}(x, ...)

\method{print}{summary.distanceFit}(x, stratum = F, ...)

\method{print}{summary.distanceList}(x, ...)

\method{summary}{distanceFit}(x, ...)

\method{summary}{distanceList}(x, ...)
}
\description{
Test
}
