% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{unmarked.wrap}
\alias{unmarked.wrap}
\title{Fit the generalized distance sampling model described by Chandler et al. (2011) with the 3 basic detection keys for a given dataset.}
\usage{
unmarked.wrap(Dataset, Trn.Value, Dist.class, Keep.class, Breaks, Trn.ID,
  Trn.length, model = NULL, Resume = F)
}
\arguments{
\item{Dataset}{Name of the input dataset.}

\item{Trn.Value}{which column in the data.frame holds the information for the densities.}

\item{Dist.class}{which column in the data.frame holds the information for the distance class.}

\item{Keep.class}{which class of observations should be used for the analysis.}

\item{Breaks}{what are the intervals used for the distance analysis.}

\item{Trn.ID}{which column in the data.frame holds the unique ID for each transect.}

\item{Trn.length}{which column in the data.frame holds the length of each transect.}

\item{model}{run only the selected model. Options are \env{"halfnorm"}, \env{"exp"}, \env{"hazard"}.}

\item{Resume}{if \env{Resume} is set to \env{TRUE} the analysis will be performed as one giant transect instead of several small transect.}
}
\value{
The function return a list containing the three models supported by the \code{\link{distsamp}} function in the package \pkg{unmarked}.
Each model composing the list is of class \code{"unmarkedFitDS"}.
}
\description{
This function allow the user to perform a Distance analysis using the unmarked package without having to reformat his data. The function also run the three different analyses by default and return all results in one list.
}
\section{Author}{
Christian Roy
}
\examples{
data(alcidae)

###Check the naïve detection histogram
hist.wrap(alcidae,Trn.Value="Count", Dist.class="Distance", Keep.class=c("A", "B", "C", "D"),
         Breaks=c(0,50,100,200,300), Trn.ID="WatchID",  Trn.length="WatchLenKm", Resume=F)

###Run the three basic unmarked models
alcidae.output<-unmarked.wrap(alcidae,Trn.Value="Count", Dist.class="Distance", Keep.class=c("A", "B", "C", "D"),
                            Breaks=c(0,50,100,200,300),  Trn.ID="WatchID",  Trn.length="WatchLenKm", Resume=T)

###Look at the structure of the output
str(alcidae.output)

####Look at the summary of each output
summary(alcidae.output[[1]])
summary(alcidae.output[[2]])
summary(alcidae.output[[3]])
#plot the resume plot for the best model
resume.plot(alcidae.output[[2]])
#END
}
\references{
S.T. Buckland, D.R. Anderson, K.P. Burnham, J.L Laake, D.L. Borchers and L. Thomas. 2001. \emph{Introduction to Distance Sampling}. Estimating abundance of biological populations. Oxford University Press.\cr
Royle, J. A., D. K. Dawson, and S. Bates. 2004. \emph{Modeling abundance effects in distance sampling}. Ecology 85:1591–1597\cr
Fiske, I. and R. B. Chandler. 2011. \emph{unmarked: An R package for fitting hierarchical models of wildlife occurrence and abundance}. Journal of Statistical Software 43:1–23.\cr
}
\seealso{
\code{\link{distsamp}}
\code{\link{distance.wrap}}
\code{\link{resume.plot}}
}

