% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{zonegulf}
\alias{zonegulf}
\title{Zones in the Gulf of St-Lawrence.}
\format{A SpatialPolygonsDataFrame.
 \describe{
   \item{\code{id}}{the global polygon to which the point is}
   \item{\code{area}}{size of the fishing zone in km}
 }}
\description{
This file gives the boundaries of fishing zones in the Gulf of St-Lawrence.
}

