% Generated by roxygen2 (4.0.2): do not edit by hand
\name{distance.wrap}
\alias{distance.wrap}
\title{Fit the distance sampling model provided by the Distance software.}
\usage{
distance.wrap(dataset, path, pathMCDS, SMP_LABEL = "SMP_LABEL",
  SMP_EFFORT = "SMP_EFFORT", DISTANCE = "DISTANCE", SIZE = "SIZE",
  STR_LABEL = "STR_LABEL", STR_AREA = "STR_AREA", breaks = c(0, 50, 100,
  200, 300), covariates = NULL, factor = NULL, lsub = NULL,
  stratum = NULL, split = TRUE, rare = NULL, period = NULL,
  detection = "All", monotone = "Strict", estimator = NULL,
  multiplier = 2, empty = NULL, verbose = FALSE)
}
\arguments{
\item{dataset}{A \code{\link{data.frame}} with observations.}

\item{path}{The path where to store the input and output files of the MCDS engine.}

\item{pathMCDS}{The path where the MCDS engine is located.}

\item{SMP_LABEL}{Name of the column to use for the transect/watch label.}

\item{SMP_EFFORT}{Length in km of the transect or the transect/watch unit.}

\item{DISTANCE}{Distance of the observation in meters.}

\item{SIZE}{Number of individuals in the observation.}

\item{STR_LABEL}{Name of the column to use as a stratum label.}

\item{STR_AREA}{Name of the column to use for the stratum area.}

\item{breaks}{A vector giving the distance intervals in meters to be used in the analysis.}

\item{covariates}{A vector giving the name of covariates.
factor A vector giving the name of factors to be used in the analysis.
lsub A named list giving the subsets to be used in the analysis. The names of the list are the names of columns used to subset the dataset.
Each element of the list is a vector giving the values to keep in the analysis for a given column. When a vector is \code{NULL},
all values are kept for this column. Default value \code{lsub = NULL}. See examples for further details.}

\item{stratum}{When \code{stratum = "STR_LABEL"}, the model will be stratified with the label.
When \code{stratum} is the name of a column in the data, the model will be
post-stratified according to this column (see Thomas et al. 2010). Default value
is \code{NULL}.}

\item{split}{When \code{split = TRUE}, separate models are ran according to the subsets determined by \code{lsub}. Default value \code{FALSE}.}

\item{rare}{This argument is used when a species has few observations to estimate a detection function.
The probability of detection is estimated from a group of similar species and a multiplier. A named list
has to be given, with the column name where species are stored as the name of the list and the name of the species as
an element (ex: \code{list(Species = "HERG")}). Default value \code{NULL}. See details.}

\item{period}{A vector of characters of length 2 containing the extreme dates for which the analysis
should be restricted. Dates have to be in the "yyyy-mm-dd" format.}

\item{detection}{Currently, set to \code{detection = "All"}. Can also be \code{detection = "Stratum"}.}

\item{monotone}{Currently, set to \code{monotone = "Strict"}.}

\item{estimator}{When set to \code{NULL}, the following key functions and expansion terms will be used: UN-CO, UN-PO, HN-CO, HN-HE, HA-CO and HA-PO.
If the user wants to choose the key functions and the expansion terms used, a list has to be given with each element a vector of
length 2 with the first element the key function and the second element the expansion term (ex: \code{list(c("HN","CO"),c("HA","PO")}).
When \code{rare != NULL}, only one set is used (UN-CO) for the final specific model. In all cases, the best model is chosen by AIC.
See details for further explanations.}

\item{multiplier}{Value by which the estimates of density or abundance are multiplied. Default value \code{multiplier = 2} meaning
only one-half of the transect is surveyed. When \code{rare != NULL}, the multiplier will be modified to account
for the probability of detection.}

\item{empty}{Determine how empty transects are to be selected in the analysis. When \code{empty = NULL}, all
empty transects are included for every element of \code{lsub}. For example, when models are splitted according
to species, empty transects and transects where a species was not detected need to be considered in the analysis for that species.
When \code{lsub} contains geographic or temporal elements, empty transects need to be restricted to the subsets given. In this case,
a vector of character has to be given with the names in \code{lsub} for which the empty transects are to be restricted. When \code{split = TRUE}
and \code{empty != NULL}, empty transects will be splitted according to the names in \code{empty}. In any case, it is assumed that
the \code{dataset} contains at least a line for every transect executed, either with or without an observation. See examples for
further details.}

\item{verbose}{When set to \code{TRUE}, prints the input file given to the MCDS engine. Default value \code{FALSE}.}
}
\value{
An object of class \code{"distanceFit"}, when \code{split = FALSE}. When \code{split = TRUE} and \code{lsub != NULL},
a named list with the different models of class \code{"distanceFit"}.
Each object of class \code{"distanceFit"} is a named list with components \code{model_fitting}, \code{parameter_estimates},
\code{chi_square_test}, \code{density_estimate}, \code{detection} and \code{path}. Elements
of the list are accessible through the \code{$} operator. For each component
except \code{detection} and \code{path}, a \code{\link{list}} of length 2 is given with
component "\code{Global}" and "\code{Stratum}". Depending on the analysis chosen,
one of these component will be empty (\code{NULL}).
}
\description{
This function fits the distance sampling model provided by the Distance program using its MCDS engine.
}
\details{
Uses the MCDS engine from program Distance. The function produces an input file and submits it to the MCDS engine
through the \code{\link{system}} function. The results are then extracted from the output files and returned as a
list object.
}
\section{Author}{
Francois Rousseu, Fran?ois Bolduc
}
\examples{
########################################
### Simple models without stratification

### Import and filter data
data(alcidae)
alcids<-filterECSAS(alcidae)

### Set arguments and parameters
path<-("c:/temp/distance")
pathMCDS<-"C:/Program Files (x86)/Distance 6"
breaks<-c(0,50,100,200,300)
STR_LABEL<-"STR_LABEL"
STR_AREA<-"STR_AREA"
SMP_LABEL<-"WatchID"
SMP_EFFORT<-"WatchLenKm"
DISTANCE<-"Distance"
SIZE<-"Count"

### Run analysis with the MCDS engine. Here, the WatchID is used as the sample.
x<-distance.wrap(alcids,path=path,pathMCDS=pathMCDS,breaks=breaks,STR_LABEL=STR_LABEL,STR_AREA=STR_AREA,SMP_LABEL=SMP_LABEL,SMP_EFFORT=SMP_EFFORT,DISTANCE=DISTANCE,SIZE=SIZE,verbose=FALSE)

### Plot results
Global.summary(model=x, file="alcidae_global", directory="C:/temp/distance")

### Run separate analysis for years 2008-2009
lsub<-list(Year=c(2008,2009))
split<-TRUE
empty<-"Year"
x<-distance.wrap(alcids,empty=empty,lsub=lsub,split=split,path=path,pathMCDS=pathMCDS,breaks=breaks,STR_LABEL=STR_LABEL,STR_AREA=STR_AREA,SMP_LABEL=SMP_LABEL,SMP_EFFORT=SMP_EFFORT,DISTANCE=DISTANCE,SIZE=SIZE,verbose=F)

### Get the names of the different models produced and plot results for the first model
names(x)
Global.summary(model=x, file="alcidae", directory="C:/temp/distance")

### Plot all results in a single line
lapply(x,resume.plot)

##############################################################
### More complex models with stratification on the Quebec data

### Import and filter data
data(quebec)
d<-filterECSAS(quebec)

### Build a shapefile with transect starts
require(rgdal)
require(sp)
transect <- data.frame(lat=d$LatStart,lon=d$LongStart)
coordinates(transect) <- ~lon + lat
transect<-SpatialPointsDataFrame(transect,data=d[,"Count",drop=FALSE])
proj4string(transect)<-CRS("+proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0")

### Build a SpatialPolygonsDataFrame from gulf zones
require(plyr)
data(zonegulf)
x<-dlply(zonegulf,.(group),function(i){Polygon(i[c(nrow(i),1:nrow(i)),1:2],hole=i$hole[1])})
x<-sapply(unique(zonegulf$id),function(i){
	temp<-x[names(x)\%in\%zonegulf$group[which(zonegulf$id==i)]]
	Polygons(temp,ID=i)
})
x<-SpatialPolygons(x,proj4string=CRS("+proj=longlat +datum=NAD83 +ellps=GRS80"))
zonegulf<-SpatialPolygonsDataFrame(x,data=data.frame(id=zonegulf$id[match(names(x),zonegulf$id)]),match.ID=FALSE)
temp<-spTransform(zonegulf,CRS("+proj=utm +zone=17 +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0"))  #project to calculate area with package rgeos
temp<-gArea(temp,byid=T)/(1000*1000)
zonegulf$area<-temp[match(zonegulf$id,names(temp))]
zonegulf<-spTransform(zonegulf,CRS(proj4string(transect)))

### Overlay transects with gulf zones
zonegulf<-spTransform(zonegulf,CRS(proj4string(transect)))
x<-over(transect,zonegulf)
d$zone<-x$id
d$zone_area<-x$area
d<-d[!is.na(d$zone),]

### Build labels for samples.
d$SMP_LABEL<-paste(d$zone,d$Date,sep="_")

### Here, every day is made of several WatchID (bouts) and the sample used will be the day (or date). Consequently, the lengths of every WatchID is summed to calculate the effort for the day.
temp<-aggregate(WatchLenKm~SMP_LABEL,data=unique(d[,c("SMP_LABEL","WatchID","WatchLenKm")]),sum)
names(temp)[2]<-"SMP_EFFORT"
d<-merge(d,temp,sort=FALSE)
d<-d[,c("zone","zone_area","Date","SMP_LABEL","SMP_EFFORT","Distance","Count","Alpha","LatStart","LongStart")]
dd<-ddply(d,.(SMP_LABEL),function(i){sum(i$Count,na.rm=TRUE)}) #eliminate duplicate lines for transect without observations
dd<-dd[dd$V1==0,] #get the label name for transect without observations
d<-d[(d$SMP_LABEL\\\%in\\\%dd$SMP_LABEL & !duplicated(d$SMP_LABEL)) | (!d$SMP_LABEL\\\%in\\\%dd$SMP_LABEL & !(d$Alpha=="")),] #keep only lines for empty transects or non-empty lines for non-empty transects
d<-d[order(d$zone),]

### Set common model arguments
path<-("c:/temp/distance")
pathMCDS<-"C:/Program Files (x86)/Distance 6"
breaks<-c(0,50,100,200,300)
SMP_LABEL<-"SMP_LABEL"
SMP_EFFORT<-"SMP_EFFORT"
DISTANCE<-"Distance"
SIZE<-"Count"

### Models with gulf zones as stratum and splitted according to species
STR_LABEL<-"zone"
STR_AREA<-"zone_area"
lsub<-list(Alpha=NULL)
split<-TRUE
stratum<-"STR_LABEL"
detection<-"All"
empty<-NULL
m<-distance.wrap(d,stratum=stratum,empty=empty,detection=detection,lsub=lsub,split=split,path=path,pathMCDS=pathMCDS,breaks=breaks,STR_LABEL=STR_LABEL,STR_AREA=STR_AREA,SMP_LABEL=SMP_LABEL,SMP_EFFORT=SMP_EFFORT,DISTANCE=DISTANCE,SIZE=SIZE,verbose=FALSE)
density.map(m,shp=zonegulf,shp.lab="id",background.shp=zonegulf,background.shp.lab="id",by.stratum=TRUE,subset="NOFU")
resume.plot(m,subset="NOFU",stratum=TRUE, file="tempo", directory="C:/temp/distance", PDF=T)

### Models splitted with gulf zones and post-stratified on species
STR_LABEL<-"zone"
STR_AREA<-"zone_area"
lsub<-list(zone=NULL)
split<-TRUE
stratum<-"Alpha"
detection<-"All"
empty<-"zone"
m<-distance.wrap(d,stratum=stratum,empty=empty,detection=detection,lsub=lsub,split=split,path=path,pathMCDS=pathMCDS,breaks=breaks,STR_LABEL=STR_LABEL,STR_AREA=STR_AREA,SMP_LABEL=SMP_LABEL,SMP_EFFORT=SMP_EFFORT,DISTANCE=DISTANCE,SIZE=SIZE,verbose=FALSE)
density.map(m,shp=zonegulf,shp.lab="id",background.shp=zonegulf,background.shp.lab="id",by.stratum=FALSE,subset="NOFU")
resume.plot(m,subset="Estuaire maritime",stratum=FALSE, file="tempo", directory="C:/temp/distance", PDF=T)

########################################################################################################
### Models with stratification on a grid with the Quebec data and splitted according to selected species

### Re-import and filter data
data(quebec)
d<-filterECSAS(quebec)
require(sp)
require(plyr)

### Build a 50000 x 50000 meters grid
size<-50000
new.grid<-create.grid(Latitude=c(44,52),Longitude=c(-70,-56),Grid.size=c(size,size),Clip=FALSE,clip.shape=canshp,projection=CRS("+proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0"))
new.grid$ID<-paste("parc",new.grid$ID,sep="")
new.grid<-new.grid[apply(gIntersects(transect,new.grid,byid=TRUE),1,any),]

### Overlay transects and grid and attribute squares to observations
x<-over(transect,new.grid)
d$square<-x$ID
d<-d[!is.na(d$square),]
d$square_area<-(size/1000)^2 #in kilometers
d$SMP_LABEL<-paste(d$square,d$Date,sep="_")

### Construct sample labels considering that day transects can overlap with multiple squares
temp<-aggregate(WatchLenKm~SMP_LABEL,data=unique(d[,c("SMP_LABEL","WatchID","WatchLenKm")]),sum)
names(temp)[2]<-"SMP_EFFORT"
d<-merge(d,temp,sort=FALSE)
d<-d[,c("square","square_area","Date","SMP_LABEL","SMP_EFFORT","Distance","Count","Alpha")]
dd<-ddply(d,.(SMP_LABEL),function(i){sum(i$Count,na.rm=TRUE)}) #eliminate duplicate lines for transect without observations
dd<-dd[dd$V1==0,] #get the label name for transect without observations
d<-d[(d$SMP_LABEL\\\%in\\\%dd$SMP_LABEL & !duplicated(d$SMP_LABEL)) | (!d$SMP_LABEL\\\%in\\\%dd$SMP_LABEL & !(d$Alpha=="")),] #keep only lines for empty transects or non-empty lines for non-empty transects
d<-d[order(d$square),]

### Set arguments and run model
STR_LABEL<-"square"
STR_AREA<-"square_area"
lsub<-list(Alpha=c("HERG","NOFU","BLKI","BOGU"))
split<-TRUE
stratum<-"STR_LABEL"
detection<-"All"
empty<-NULL
m<-distance.wrap(d,stratum=stratum,empty=empty,detection=detection,lsub=lsub,split=split,path=path,pathMCDS=pathMCDS,breaks=breaks,STR_LABEL=STR_LABEL,STR_AREA=STR_AREA,SMP_LABEL=SMP_LABEL,SMP_EFFORT=SMP_EFFORT,DISTANCE=DISTANCE,SIZE=SIZE,verbose=FALSE)
density.map(m,shp=new.grid,shp.lab="ID",background.shp=zonegulf,background.shp.lab="id",by.stratum=TRUE,subset="NOFU")
resume.plot(m,subset="NOFU",stratum=TRUE, file="tempo", directory="C:/temp/distance", PDF=T)

### Models for two species and restricted to the year 2008 and 2009
d$Year<-as.numeric(substr(d$Date,1,4))
lsub<-list(Alpha=c("RTLO","COLO"),Year=2008:2009)
split<-TRUE
empty<-"Year"
m<-distance.wrap(d,stratum=stratum,empty=empty,detection=detection,lsub=lsub,split=split,path=path,pathMCDS=pathMCDS,breaks=breaks,STR_LABEL=STR_LABEL,STR_AREA=STR_AREA,SMP_LABEL=SMP_LABEL,SMP_EFFORT=SMP_EFFORT,DISTANCE=DISTANCE,SIZE=SIZE,verbose=FALSE)
density.map(m,shp=new.grid,shp.lab="ID",background.shp=zonegulf,background.shp.lab="id",by.stratum=TRUE,subset="2008_RTLO")
#################################
### Examples with other arguments
### A model stratified by species and restricted to a period  for which we want the a HN and HA key function with cosine expansion
DISTANCE<-"Distance"
SIZE<-"Count"
period<-c("2009-05-01","2009-11-30")
stratum<-"Alpha"
estimator<-list(c("HN","CO"),c("HA","CO"))
m<-distance.wrap(d,path=path,pathMCDS=pathMCDS,DISTANCE=DISTANCE,SIZE=SIZE,stratum=stratum,estimator=estimator,period=period)
resume.plot(m,stratum=TRUE, file="tempo", directory="C:/temp/distance", PDF=T)

### A model estimated for a selected rarer species for which the probability of detection is estimated through other similar species and making use of multipliers
DISTANCE<-"Distance"
SIZE<-"Count"
lsub<-list(Alpha=c("HERG","GBBG","BLKI","RBGU","BOGU"))
rare<-list(Alpha="BOGU")
m<-distance.wrap(d,path=path,pathMCDS=pathMCDS,DISTANCE=DISTANCE,SIZE=SIZE,lsub=lsub,rare=rare)
resume.plot(m,stratum=TRUE, file="tempo", directory="C:/temp/distance", PDF=T)

#END
}
\references{
Thomas, L., S.T. Buckland, E.A. Rexstad, J. L. Laake, S. Strindberg, S. L. Hedley, J. R.B. Bishop,
T. A. Marques, and K. P. Burnham. 2010. \emph{Distance software: design and analysis of distance sampling
surveys for estimating population size.} Journal of Applied Ecology 47: 5-14.
DOI: 10.1111/j.1365-2664.2009.01737.x
}

