% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2MCDS-package.r
\docType{package}
\name{R2MCDS}
\alias{R2MCDS}
\alias{R2MCDS-package}
\title{Calling  Distance 6.2 MCDS engine from R.}
\description{
This package will allow the user to call the MCDS engine from Distance 6.2 in the R environment.
}
\details{
\tabular{ll}{
Package: \tab GeoAviR\cr
Type: \tab Package\cr
Version: \tab 0.8.3\cr
Date: \tab 2015-10-26\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
}
\examples{
### Import and filter data
data(alcidae)
alcids<-filterECSAS(alcidae)
##END
}
\references{
S.T. Buckland, D.R. Anderson, K.P. Burnham, J.L Laake, D.L. Borchers and L. Thomas. 2001. \emph{Introduction to Distance Sampling}. Estimating abundance of biological populations. Oxford University Press.\cr
Royle, J. A., D. K. Dawson, and S. Bates. 2004. \emph{Modeling abundance effects in distance sampling}. Ecology 85:1591-1597\cr
Fiske, I. and R. B. Chandler. 2011. \emph{unmarked: An R package for fitting hierarchical models of wildlife occurrence and abundance}. Journal of Statistical Software 43:1-23.\cr 
Thomas, L., S.T. Buckland, E.A. Rexstad, J. L. Laake, S. Strindberg, S. L. Hedley, J. R.B. Bishop, T. A. Marques, and K. P. Burnham. 2010.  Distance software: design and analysis of distance sampling surveys for estimating population size.  Journal of Applied Ecology 47: 5-14.
}
