% Generated by roxygen2 (4.0.2): do not edit by hand
\name{strip.wrap}
\alias{strip.wrap}
\title{Fit the distance sampling model provided by the Distance software.}
\usage{
strip.wrap(dataset, path, pathMCDS, SMP_LABEL = "SMP_LABEL",
  SMP_EFFORT = "SMP_EFFORT", DISTANCE = "DISTANCE", SIZE = "SIZE",
  STR_LABEL = "STR_LABEL", STR_AREA = "STR_AREA", breaks = c(0, 50, 100,
  200, 300), lsub = NULL, stratum = NULL, split = TRUE, period = NULL,
  detection = "All", multiplier = 2, empty = NULL, verbose = FALSE)
}
\arguments{
\item{dataset}{A \code{\link{data.frame}} with observations.}

\item{path}{The path where to store the input and output files of the MCDS engine.}

\item{pathMCDS}{The path where the MCDS engine is located.}

\item{SMP_LABEL}{Name of the column to use for the transect/watch label.}

\item{SMP_EFFORT}{Length in km of the transect or the transect/watch unit.}

\item{DISTANCE}{Distance of the observation in meters.}

\item{SIZE}{Number of individuals in the observation.}

\item{STR_LABEL}{Name of the column to use as a stratum label.}

\item{STR_AREA}{Name of the column to use for the stratum area.}

\item{breaks}{A vector giving the distance intervals in meters to be used in the analysis.}

\item{stratum}{When \code{stratum = "STR_LABEL"}, the model will be stratified with the label.
When \code{stratum} is the name of a column in the data, the model will be
post-stratified according to this column (see Thomas et al. 2010). Default value
is \code{NULL}.}

\item{split}{When \code{split = TRUE}, separate models are ran according to the subsets determined by \code{lsub}. Default value \code{FALSE}.}

\item{period}{A vector of characters of length 2 containing the extreme dates for which the analysis
should be restricted. Dates have to be in the "yyyy-mm-dd" format.}

\item{multiplier}{Value by which the estimates of density or abundance are multiplied. Default value \code{multiplier = 2} meaning
only one-half of the transect is surveyed. When \code{rare != NULL}, the multiplier will be modified to account
for the probability of detection.}

\item{empty}{Determine how empty transects are to be selected in the analysis. When \code{empty = NULL}, all
empty transects are included for every element of \code{lsub}. For example, when models are splitted according
to species, empty transects and transects where a species was not detected need to be considered in the analysis for that species.
When \code{lsub} contains geographic or temporal elements, empty transects need to be restricted to the subsets given. In this case,
a vector of character has to be given with the names in \code{lsub} for which the empty transects are to be restricted. When \code{split = TRUE}
and \code{empty != NULL}, empty transects will be splitted according to the names in \code{empty}. In any case, it is assumed that
the \code{dataset} contains at least a line for every transect executed, either with or without an observation. See examples for
further details.}

\item{verbose}{When set to \code{TRUE}, prints the input file given to the MCDS engine. Default value \code{FALSE}.}
}
\value{
An object of class \code{"distanceFit"}, when \code{split = FALSE}. When \code{split = TRUE} and \code{lsub != NULL},
a named list with the different models of class \code{"distanceFit"}.
Each object of class \code{"distanceFit"} is a named list with components \code{model_fitting}, \code{parameter_estimates},
\code{chi_square_test}, \code{density_estimate}, \code{detection} and \code{path}. Elements
of the list are accessible through the \code{$} operator. For each component
except \code{detection} and \code{path}, a \code{\link{list}} of length 2 is given with
component "\code{Global}" and "\code{Stratum}". Depending on the analysis chosen,
one of these component will be empty (\code{NULL}).
}
\description{
This function use the  MCDS engine of the Distance program to fit a uniform detection function.
}
\details{
Uses the MCDS engine from program Distance. The function produces an input file and submits it to the MCDS engine
through the \code{\link{system}} function. The results are then extracted from the output files and returned as a
list object.
}
\section{Author}{
Francois Rousseu, Christian Roy, François Bolduc
}
\references{
Thomas, L., S.T. Buckland, E.A. Rexstad, J. L. Laake, S. Strindberg, S. L. Hedley, J. R.B. Bishop,
T. A. Marques, and K. P. Burnham. 2010. \emph{Distance software: design and analysis of distance sampling
surveys for estimating population size.} Journal of Applied Ecology 47: 5-14.
DOI: 10.1111/j.1365-2664.2009.01737.x
}

